AI_Employee_Vault/mcp_servers/email_mcp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
pydantic==2.5.0
python-dotenv==1.0.0
